refactor: fixing mappers and register entry

+++ src/main/java/org/sistema/arroz/riceservice/modules/merchandiseEntry/application/port/in/MerchandiseEntryDetailToRegister.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class MerchandiseEntryDetailToRegister {
     Long supplyId;
-    Integer entryCant;
+    Double entryCant;
 }
